pyproject.toml
[project]
name = "spring-job-api-v2"
version = "0.1.0"
description = "CV Agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.121.3",
    "langchain>=1.0.8",
    "langchain-core>=1.0.7",
    "langchain-google-genai>=3.1.0",
    "langchain-mcp-adapters>=0.1.14",
    "langgraph>=1.0.3",
    "langgraph-cli[inmem]>=0.4.7",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cv_agent"]
